Temperature detail table, field float, navigation admission to temperature record

+++ database/migrations/2024_12_10_020436_create_temperature_details_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('temperature_record_id')
                 ->constrained()->onDelete('restrict');
             $table->foreignId('nurse_id')->constrained(table:'users')->onDelete('restrict');
-            $table->integer('temperature');
+            $table->float('temperature');
             $table->integer('evacuations');
             $table->string('urinations', 2);
             $table->boolean('active')->default(true);
